% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{circular_variable_difference}
\alias{circular_variable_difference}
\title{A function which calculates the smallest difference between two circularly defined values.}
\usage{
circular_variable_difference(value1,value2,values_in_full_circle)
}
\arguments{
\item{value1}{value which is compared against the other value}

\item{value2}{value which the first value is compared against}

\item{values_in_full_circle}{maximum value for value1 and value2}
}
\value{
smallest difference for the value1 compared against value2
}
\description{
This function calculates the smallest difference between two circularly defined values. Examples from circular variables are (azimuth angle, hour of the day, day of the year)
}
\details{
Function call (circular_variable_difference(10,350,360)) gives 20 degrees as the result: 10 degrees is 20 degrees more compared to 350 degrees, not 340 degrees less.
}
