% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{conditional.quantile.extremes}
\alias{conditional.quantile.extremes}
\title{Plots conditional quantile plots from two datasets}
\usage{
function(pred, obs, bins = NULL, thrs = c(10,20), main, na.rm=TRUE)
}
\arguments{
\item{pred}{Forecasted value. ([n,1] vector, n = No. of forecasts)}

\item{obs}{Observed value.([n,1] vector, n = No. of observations)}

\item{bins}{Bins for forecast and observed values. A minimum number of values are required to calculate meaningful statistics. So for variables that are continuous, such as temperature, it is frequently convenient to bin these values. ([m,1] vector, m = No. of bins)}

\item{thrs}{The minimum number of values in a bin required to calculate the 25th and 75th quantiles and the 10th and 90th percentiles respectively. The median+min+max values will always be displayed. ( [2,1] vector)}

\item{main}{Plot title}

\item{...}{plotting options}
}
\value{
A plot
}
\description{
This function plots quantile-quantile plot from two datasets. This is an improved version from conditional.quantile() -function in "verification"-package
}
\details{
add here
}
