% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{define_time_series}
\alias{define_time_series}
\title{Defining time series with fixed intervals}
\usage{
define_time_series(begin_date, end_date, interval_in_hours)
}
\arguments{
\item{begin_date}{First date of the time series}

\item{end_date}{All time stamps in the time series must be smaller than this last value}

\item{interval_in_hours}{Interval of two consecutive time stamps in the time series}
}
\value{
A regularly spaced time series
}
\description{
Define a regularly spaced time series
}
\details{
Use POSIXct-type date-time stamps! Time series are rounded to nearest even 3-hour ({00,03,06,09,12,15,18,21}). Default values of the parameters correspond to dates used in MOS training (begin_date = "2011-12-01 00:00:00 GMT", end_date = Sys.time()).
}
