% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Test if x not in y}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{object which is matched in y}

\item{y}{the list to what x is matched against}
}
\value{
A vector of the same length as x. A logical vector, indicating if a match was NOT located for each element of x: thus the values are TRUE or FALSE and never NA.
}
\description{
This is the opposite function to the \pkg{base} inline function \%in\%
}
\details{
\pkg{base} inline function \%in\% is in reality written as \%in\% <- function(x, table) match(x, table, nomatch = 0) > 0,
\%notin\% is written simply as \%notin\% <- function(x,y) !(x \%in\% y)
}
