% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{\%||\%}
\alias{\%||\%}
\title{If x is null then return y}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Object which may return NULL (most likely a function)}

\item{y}{Default value in case of NULL}
}
\value{
Either the value returned by x or default value y
}
\description{
This is an inline function that provides a default value in cases when the result is NULL
}
\details{
The whole code of this function is `\%||\%` <- function(x, y) if (!is.null(x)) x else y \cr
Typical use for this inline function is to provide a default value if a function gives NULL, e.g. function_that_might_return_null() \%||\% default value
}
\seealso{
\url{http://adv-r.had.co.nz/Functions.html}
}
