% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{interpolate_NA_values}
\alias{interpolate_NA_values}
\title{Interpolating (timeseries) values}
\usage{
Interpolate_timeseries(data_to_be_interpolated, interpolation_method, maximum_gap, ...)
}
\arguments{
\item{data_to_be_interpolated}{1-dimensional data with NA values.}

\item{interpolation_method}{The selected interpolation method. ("repeat_previous","linear_interp","spline_interp","no_interp") are supported.}

\item{maxgap}{Maximum gap (number of consecutive NA-values) allowed, which still is being interpolated}

\item{...}{Possible additional parameters passed to zoo-package interpolation functions}
}
\value{
Interpolated time series of same size as the input data.
}
\description{
This is a generic NA-interpolation switch wrapper which allows the use of several interpolation methods
}
\details{
data_to_be_interpolated is 1-dimensional (without actual time stamps). It contains NA-values for the missing time stamps. Returned data is the same size but with interpolated values.
}
