% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{is.NullOb}
\alias{is.NullOb}
\title{A function that tests whether an object is either NULL _or_ a list of NULLs}
\usage{
is.NullOb(list)
}
\arguments{
\item{list}{list which is checked}
}
\value{
A logical list of same size as input list
}
\description{
This is a small function that tests whether all elements in a list are either NULL values OR a list of NULL values
}
