% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{lappend}
\alias{lappend}
\title{Add object to a list}
\usage{
lappend(lst, obj, listname)
}
\arguments{
\item{lst}{a list to which object is added}

\item{obj}{an object which is added}

\item{listname}{name of the added object as it appers in the list}
}
\value{
the original list with the object added
}
\description{
This function adds objects to the end of a list
}
\details{
As the list can contain miscellaneous variable types, this function should also work regardless of the variable type. No thorough tests are done, however.
}
\examples{
all_station_lists <- lappend(all_station_lists,loaded_list$wmon,
"mos_stations_homogeneous_Europe")
}
