% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{loadRData}
\alias{loadRData}
\title{A function for reading in data files}
\usage{
loadRData(fileName)
}
\arguments{
\item{fileName}{The file which is loaded}
}
\value{
the data stored in .RData file
}
\description{
This function reads in RData files and returns the objects in it to the environment where the function was called from. Using this function the data can be assigned to an arbitrary variable name instead of that used in the RData file itself while it was saved.
}
