% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{MOSpointutils-package}
\alias{MOSpointutils}
\alias{MOSpointutils-package}
\title{Miscellaneous tools mainly for data handling and manipulation}
\description{
These functions make life a bit easier when doing analysis and training for MOS point data
}
