% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_points_on_map}
\alias{plot_points_on_map}
\title{Plots points to a map on a user defined lat-lon box, colorscale, limits and legend location}
\usage{
function(plot_data,plot_data_area,breaks,colorscale,point_size,plotting_title)
}
\arguments{
\item{plot_data}{a data matrix (first column data, second lon, third lat)}

\item{plot_data_area}{defines the data points which constitute the plotting area}

\item{breaks}{a vector defining breaks for plotting}

\item{colorscale}{the colorscale that is used in plot (if colorscale==points, do not plot any scale)}

\item{point_size}{size of point to be plotted}

\item{plotting_title}{plotting title}
}
\value{
A plot
}
\description{
This function plots points to a geographical map, with points having values which are indicated by a colorscale
}
\details{
add here
}
