% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{rmNullObs}
\alias{rmNullObs}
\title{A function that recursively removes all only-NULL lists among nested lists}
\usage{
rmNullObs(list)
}
\arguments{
\item{list}{list where NULL values are removed}
}
\value{
A list without NULL values
}
\description{
This function removes all NULL lists from the input data. This function is recursive, so the input list can have as much nested lists as possible.
}
