% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_useful.R
\name{unwhich}
\alias{unwhich}
\title{Converting integers to a boolean list}
\usage{
unwhich(x,n)
}
\arguments{
\item{x}{Integers that indicate TRUE values}

\item{n}{Boolean list length}
}
\value{
Boolean list with length n
}
\description{
Which (in \pkg{base}) converts true/false values to integers, unwhich does the opposite (n corresponds to vector length where true values are assigned to)
}
\details{
The function first creates n-length list of FALSE values. After this, indices x are replaced with TRUE values.
}
\seealso{
\url{http://adv-r.had.co.nz/Subsetting.html}
}
